pyproject.toml
[project]
name = "Arke"
version = "0.1.0"
description = "A fast, efficient Retrieval-Augmented Generation (RAG) system built with modern AI tools for document querying and knowledge base management."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-redis>=0.2.5",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "redis>=7.1.0",
    "python-magic>=0.4.27",
    "kreuzberg>=4.0.0",
    "rich>=14.2.0",
    "openai>=2.14.0",
    "sentence-transformers>=5.2.0",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.24.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "langgraph-checkpoint-redis>=0.3.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
